+++ airbyte-integrations/bases/source-acceptance-test/unit_tests/test_plugin.py
@@ -4,9 +4,8 @@
 import pytest
 from source_acceptance_test import config, plugin
 
-STRICT_TEST_MODE = config.Config.TestMode.strict
-MEDIUM_TEST_MODE = config.Config.TestMode.medium
-LIGHT_TEST_MODE = config.Config.TestMode.light
+HIGH_TEST_STRICTNESS_LEVEL = config.Config.TestStrictnessLevel.high
+LOW_TEST_STRICTNESS_LEVEL = config.Config.TestStrictnessLevel.low
 
 PARAMETRIZE_ACTION = plugin.TestAction.PARAMETRIZE
 SKIP_ACTION = plugin.TestAction.SKIP
@@ -47,103 +46,76 @@ def test_pytest_generate_tests(mocker, parametrize_skip_or_fail_return):
 
 
 @pytest.mark.parametrize(
-    "TestClass, test_class_mandatory_for_test_modes, global_test_mode, test_configuration, expected_action, expected_reason",
+    "TestClass, test_class_MANDATORY_FOR_TEST_STRICTNESS_LEVELS, global_test_mode, test_configuration, expected_action, expected_reason",
     [
         pytest.param(
             MyTestClass,
-            (STRICT_TEST_MODE),
-            STRICT_TEST_MODE,
+            (HIGH_TEST_STRICTNESS_LEVEL),
+            HIGH_TEST_STRICTNESS_LEVEL,
             None,
             FAIL_ACTION,
-            "MyTestClass.dumb_test_function failed: it was not configured but must be according to the current strict test mode.",
-            id="Discovered test is mandatory in strict mode, we're in strict mode, it was not configured: FAIL",
+            "MyTestClass.dumb_test_function failed: it was not configured but must be according to the current high test strictness level.",
+            id="Discovered test is mandatory in high test strictness level, we're in high test strictness level, it was not configured: FAIL",
         ),
         pytest.param(
             MyTestClass,
-            (STRICT_TEST_MODE),
-            LIGHT_TEST_MODE,
+            (HIGH_TEST_STRICTNESS_LEVEL),
+            LOW_TEST_STRICTNESS_LEVEL,
             None,
             SKIP_ACTION,
             "Skipping MyTestClass.dumb_test_function: not found in the config.",
-            id="Discovered test is mandatory in strict mode, we are in light mode, it is not configured: SKIP",
+            id="Discovered test is mandatory in high test strictness level, we are in low strictness level, it is not configured: SKIP",
         ),
         pytest.param(
             MyTestClass,
             set(),
-            STRICT_TEST_MODE,
+            HIGH_TEST_STRICTNESS_LEVEL,
             None,
             SKIP_ACTION,
             "Skipping MyTestClass.dumb_test_function: not found in the config.",
-            id="Discovered test is not mandatory in any mode, it was not configured: SKIP",
+            id="Discovered test is not mandatory in any test strictness level, it was not configured: SKIP",
         ),
         pytest.param(
             MyTestClass,
-            (STRICT_TEST_MODE),
-            STRICT_TEST_MODE,
+            (HIGH_TEST_STRICTNESS_LEVEL),
+            HIGH_TEST_STRICTNESS_LEVEL,
             config.GenericTestConfig(bypass_reason="A good reason."),
             SKIP_ACTION,
             "Skipping MyTestClass.dumb_test_function: A good reason.",
-            id="Discovered test is mandatory in strict mode, a bypass reason was provided: SKIP",
+            id="Discovered test is mandatory in high test strictness level, a bypass reason was provided: SKIP",
         ),
         pytest.param(
             MyTestClass,
-            (STRICT_TEST_MODE),
-            LIGHT_TEST_MODE,
+            (HIGH_TEST_STRICTNESS_LEVEL),
+            LOW_TEST_STRICTNESS_LEVEL,
             config.GenericTestConfig(bypass_reason="A good reason."),
             SKIP_ACTION,
             "Skipping MyTestClass.dumb_test_function: A good reason.",
-            id="Discovered test is mandatory in strict mode, we are in light mode, a bypass reason was provided: SKIP (with bypass reason shown)",
+            id="Discovered test is mandatory in high test strictness level, we are in low test strictness level, a bypass reason was provided: SKIP (with bypass reason shown)",
         ),
         pytest.param(
             MyTestClass,
-            (MEDIUM_TEST_MODE, STRICT_TEST_MODE),
-            LIGHT_TEST_MODE,
-            None,
-            SKIP_ACTION,
-            "Skipping MyTestClass.dumb_test_function: not found in the config.",
-            id="Discovered test is mandatory in medium and strict mode, we're in light mode, it was not configured: SKIP",
-        ),
-        pytest.param(
-            MyTestClass,
-            (STRICT_TEST_MODE),
-            STRICT_TEST_MODE,
+            (HIGH_TEST_STRICTNESS_LEVEL),
+            HIGH_TEST_STRICTNESS_LEVEL,
             config.GenericTestConfig(tests=[config.SpecTestConfig()]),
             PARAMETRIZE_ACTION,
             "Parametrize MyTestClass.dumb_test_function: tests are configured.",
-            id="[Strict mode] Discovered test is configured: PARAMETRIZE",
+            id="[High test strictness level] Discovered test is configured: PARAMETRIZE",
         ),
         pytest.param(
             MyTestClass,
-            (STRICT_TEST_MODE),
-            LIGHT_TEST_MODE,
+            (HIGH_TEST_STRICTNESS_LEVEL),
+            LOW_TEST_STRICTNESS_LEVEL,
             config.GenericTestConfig(tests=[config.SpecTestConfig()]),
             PARAMETRIZE_ACTION,
             "Parametrize MyTestClass.dumb_test_function: tests are configured.",
-            id="[Light mode] Discovered test is configured: PARAMETRIZE",
-        ),
-        pytest.param(
-            MyTestClass,
-            (STRICT_TEST_MODE),
-            MEDIUM_TEST_MODE,
-            config.GenericTestConfig(tests=[config.SpecTestConfig()]),
-            PARAMETRIZE_ACTION,
-            "Parametrize MyTestClass.dumb_test_function: tests are configured.",
-            id="[Medium mode] Discovered test is configured: PARAMETRIZE",
-        ),
-        pytest.param(
-            MyTestClass,
-            set(),
-            LIGHT_TEST_MODE,
-            config.GenericTestConfig(tests=[config.SpecTestConfig()]),
-            PARAMETRIZE_ACTION,
-            "Parametrize MyTestClass.dumb_test_function: tests are configured.",
-            id="[Light mode] Discovered test is configured, test is not mandatory in any mode: PARAMETRIZE",
+            id="[Low test strictness level] Discovered test is configured: PARAMETRIZE",
         ),
     ],
 )
 def test_parametrize_skip_or_fail(
-    TestClass, test_class_mandatory_for_test_modes, global_test_mode, test_configuration, expected_action, expected_reason
+    TestClass, test_class_MANDATORY_FOR_TEST_STRICTNESS_LEVELS, global_test_mode, test_configuration, expected_action, expected_reason
 ):
-    TestClass.MANDATORY_FOR_TEST_MODES = test_class_mandatory_for_test_modes
+    TestClass.MANDATORY_FOR_TEST_STRICTNESS_LEVELS = test_class_MANDATORY_FOR_TEST_STRICTNESS_LEVELS
     test_action, reason = plugin.parametrize_skip_or_fail(TestClass, TestClass.dumb_test_function, global_test_mode, test_configuration)
     assert (test_action, reason) == (expected_action, expected_reason)